backend/requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
matplotlib
pillow
python-multipart
pytest
